Add clone method to logger
Add clone method to logger.

Create base logger class implementing clone.

Add generic to ILogger which contains the properties stored in the class.

Add unit tests to verify deep and shallow cloning.

Update license to GPL3.

Notes:

Input properties are stored in a protected variable used in the clone method
to replicate itself. To make this typesafe I updated ILogger to be generic,
where the type is the input properties of the logger.

The clone method accepts overrides supporting all of the arguments of the
original constructor, creating a new instance of the logger. If no overrides
are set, the new logger instance will have the same configuration as the parent.